gcf/requirements.txt
google-cloud-storage
google-cloud-aiplatform
pymysql